pyproject.toml
[project]
name = "ai-agent-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "beautifulsoup4>=4.14.3",
    "datasets>=4.4.2",
    "ddgs>=9.10.0",
    "deepagents>=0.3.2",
    "docling[vlm]>=2.66.0",
    "docutils>=0.22.4",
    "docx2txt>=0.9",
    "duckduckgo-search>=8.1.1",
    "e2b-code-interpreter>=2.4.1",
    "email-validator>=2.3.0",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "fastembed>=0.7.3",
    "fastembed-gpu>=0.7.3",
    "google-genai>=1.57.0",
    "hf-xet>=1.2.0",
    "huggingface-hub>=0.36.0",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "langchain>=1.2.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.3",
    "langchain-google-vertexai>=3.2.1",
    "langchain-huggingface>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.2",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "langsmith>=0.6.1",
    "lxml>=6.0.2",
    "markitdown[all]>=0.1.4",
    "matplotlib>=3.10.8",
    "nest-asyncio>=1.6.0",
    "nltk>=3.9.2",
    "numba>=0.62.1",
    "numpy>=2.2.6",
    "ollama>=0.6.1",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pymupdf4llm>=0.2.8",
    "pymysql>=1.1.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-pptx>=1.0.2",
    "pytube>=15.0.0",
    "rank-bm25>=0.2.2",
    "selenium>=4.39.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "tavily-python>=0.7.17",
    "textblob>=0.19.0",
    "tiktoken>=0.12.0",
    "tokenizers>=0.22.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.3",
    "typing-extensions>=4.15.0",
    "unstructured>=0.18.26",
    "wikipedia>=1.4.0",
    "xmltodict>=1.0.2",
    "youtube-transcript-api>=1.0.3",
]
